Dowload and parse list C#
Added code to download and parse the app list and then list the apps it found in a data grid view with a checkbox. I need to work on getting the install button working with more logging. Also need to add more programs to the app list because right now its just discord.